test: registerDefaultUser moved to Utils
